Improve code quality, monitoring, CLI features, and edge case handling

+++ app/services/Vlans_Services.py
@@ -1,10 +1,12 @@
-import re, os, httpx
+import re
+import os
+import httpx
 from fastapi import HTTPException
 from dotenv import load_dotenv
 from email.utils import formatdate
-from app.models.Vlan import Vlan_Post_Request, VlanWrapper, SonicVLAN, SonicVLANMember
-
-
+from app.models.Vlan import Vlan_Post_Request, VlanWrapper, SonicVLAN, SonicVLANMember, Vlan_Get_Response
+from app.services.Port_Op_Services import get_po_service
+from app.models.Port import Port_Oper_Response
 
 load_dotenv()
  
@@ -16,17 +18,48 @@ RESTCONF_HEADERS = {
     "Content-Type": "application/yang-data+json"
 }
 
-ETH_INTERFACES = {
-    "Ethernet0", "Ethernet4", "Ethernet8", "Ethernet12", "Ethernet16", "Ethernet20",
-    "Ethernet24", "Ethernet28", "Ethernet32", "Ethernet36", "Ethernet40", "Ethernet44",
-    "Ethernet48", "Ethernet52", "Ethernet56", "Ethernet60", "Ethernet64", "Ethernet68",
-    "Ethernet72", "Ethernet76", "Ethernet80", "Ethernet84", "Ethernet88", "Ethernet92",
-    "Ethernet96", "Ethernet100", "Ethernet104", "Ethernet108", "Ethernet112", "Ethernet116",
-    "Ethernet120", "Ethernet124"
-}
 
 
-def validate_vlan_data(vlan_list:SonicVLAN, member_list: SonicVLANMember):
+async def get_Ethernet_List():
+    json_Port_data = await get_po_service()
+    response = Port_Oper_Response(**json_Port_data)
+    Ethernets = []
+    ports_list = response.port.PORT_TABLE.PORT_TABLE_LIST
+    for port in ports_list:
+        Ethernets.append(port.ifname)
+
+    Ethernets.sort(key=lambda x: int(x.replace("Ethernet", "")))
+    return Ethernets
+
+
+async def check_untagged_if(eth:str):
+    json_Vlan_data = await fetch_vlans()
+    if not json_Vlan_data: # if not vlans exist so response will be empty
+        return
+    response = Vlan_Get_Response(**json_Vlan_data)
+    members_list = response.wrapper.VLAN_MEMBER.VLAN_MEMBER_LIST
+    if members_list is not None:
+        for i in members_list:
+            if i.ifname == eth and i.tagging_mode == "untagged":
+                return True
+    return False
+
+
+async def check_Vlan_exist(vlan:str):
+    json_Vlan_data = await fetch_vlans()
+    if not json_Vlan_data: # if not vlans exist so response will be empty
+        return
+    response = Vlan_Get_Response(**json_Vlan_data)
+    vlan_names = response.wrapper.VLAN.VLAN_LIST
+    for name in vlan_names:
+        if name == vlan:
+            return True
+    return False
+
+
+async def validate_vlan_data(vlan_list:SonicVLAN, member_list: SonicVLANMember):
+    
+    ETH_INTERFACES = await get_Ethernet_List()
     try:
     
         for vlan in vlan_list:
@@ -50,15 +83,20 @@ def validate_vlan_data(vlan_list:SonicVLAN, member_list: SonicVLANMember):
                 member_vlan_name=member.name
 
                 if re.fullmatch(r'Vlan\d+', member_vlan_name) is None:
-                    raise ValueError(f"VLAN name must match 'Vlan<id>' pattern")
+                    raise ValueError("VLAN name must match 'Vlan<id>' pattern")
                 
                 ifname = member.ifname
                 if ifname not in ETH_INTERFACES:
                     raise ValueError(f"Invalid interface name '{ifname}'. Allowed: {', '.join(ETH_INTERFACES)}")
-
+                
                 if member.tagging_mode not in {"tagged", "untagged"}:
-                    raise ValueError(f"Invalid tagging_mode. Must be 'tagged' or 'untagged'.")
-
+                    raise ValueError("Invalid tagging_mode, Must be 'tagged' or 'untagged'.")
+                
+                if member.tagging_mode == "untagged":
+                    is_untagged = await check_untagged_if(ifname)
+                    if is_untagged:
+                        raise ValueError(f"Interface '{ifname}' is already configured as untagged in another VLAN. Choose another interface or change the tagging mode.")
+    
     except ValueError as ve:
         raise HTTPException(status_code=400, detail=str(ve))
     
@@ -73,16 +111,18 @@ async def fetch_vlans():
 
             response.raise_for_status()
             return response.json()
-    except Exception as e:
-        raise HTTPException(status_code=500, detail=str(e))
+        
+    except httpx.HTTPStatusError as e:
+        raise HTTPException(status_code=e.response.status_code, detail=str(e))  
+    
 
 
 async def post_vlans_service(request:Vlan_Post_Request):
 
-
     vlan_List= request.vlan.VLAN_LIST
     member_List= request.members.VLAN_MEMBER_LIST
-    validate_vlan_data(vlan_List, member_List)
+    await validate_vlan_data(vlan_List, member_List)
+
     try:
         
         async with httpx.AsyncClient(verify=False, timeout = 10.0) as client:
@@ -98,20 +138,24 @@ async def post_vlans_service(request:Vlan_Post_Request):
             }
 
         
-    except Exception as e:
-        raise HTTPException(status_code=500, detail=str(e))
+    except httpx.HTTPStatusError as e:
+        raise HTTPException(status_code=e.response.status_code, detail=str(e))  
     
 
 
-async def put_vlan_service(payload: dict):
-    url = f"{SONIC_BASE_URL}/restconf/data/sonic-vlan:sonic-vlan"
+async def put_vlan_service(request:VlanWrapper):
+    
+    vlan_List= request.wrapper.VLAN.VLAN_LIST
+    member_List= request.wrapper.VLAN_MEMBER.VLAN_MEMBER_LIST
+    await validate_vlan_data(vlan_List, member_List)
+
     try:
 
         async with httpx.AsyncClient(verify=False, timeout = 10.0) as client:
             response = await client.put(
-                url,
+                f"{SONIC_BASE_URL}/restconf/data/sonic-vlan:sonic-vlan",
                 headers=RESTCONF_HEADERS,
-                json=payload  # ← raw body passed directly)
+                json=request.model_dump(by_alias=True)
             )
 
             response.raise_for_status()
@@ -120,16 +164,16 @@ async def put_vlan_service(payload: dict):
                 "message": "VLAN configuration updated",
                 "date": formatdate(timeval=None, usegmt=True)
             }
-    except Exception as e: 
-        raise HTTPException(status_code=500, detail=str(e))
-
+    
+    except httpx.HTTPStatusError as e:
+        raise HTTPException(status_code=e.response.status_code, detail=str(e))  
 
 
 async def patch_vlans_service(request:VlanWrapper):
     
-    vlan_List= request.wrapper.vlan.VLAN_LIST
-    member_List= request.wrapper.members.VLAN_MEMBER_LIST
-    validate_vlan_data(vlan_List, member_List)
+    vlan_List= request.wrapper.VLAN.VLAN_LIST
+    member_List= request.wrapper.VLAN_MEMBER.VLAN_MEMBER_LIST
+    await validate_vlan_data(vlan_List, member_List)
 
     try:
 
@@ -147,51 +191,60 @@ async def patch_vlans_service(request:VlanWrapper):
                 "date": formatdate(timeval=None, usegmt=True)
             }
 
-    except Exception as e:
-        raise HTTPException(status_code=500, detail=str(e))
+    except httpx.HTTPStatusError as e:
+        raise HTTPException(status_code=e.response.status_code, detail=str(e))  
 
 
 async def delete_all_vlans_from_switch():
+
+    vlans = await fetch_vlans()
+    if not vlans:
+        raise HTTPException(status_code=404, detail="No VLANs found in the switch")
     try:
         async with httpx.AsyncClient(verify=False) as client:
             response = await client.delete(f"{SONIC_BASE_URL}/restconf/data/sonic-vlan:sonic-vlan", headers=RESTCONF_HEADERS, timeout=10.0)
 
             response.raise_for_status()
             return {"detail": "All VLANs deleted from switch successfully."}
-    except Exception as e:
-            raise HTTPException(
-                status_code=500,
-                detail=str(e)
-            )
+
+    except httpx.HTTPStatusError as e:
+        raise HTTPException(status_code=e.response.status_code, detail=str(e))  
 
     
     
 async def delete_vlan_by_name(vlan_name: str):
     vlan_url = f"{SONIC_BASE_URL}/restconf/data/sonic-vlan:sonic-vlan/VLAN/VLAN_LIST={vlan_name}"
+    
+    exist = await check_Vlan_exist(vlan_name)
+    if not exist:
+        raise HTTPException(status_code=404, detail="Vlan name not found")
+    
     try:
         async with httpx.AsyncClient(verify=False) as client:
             response = await client.delete(vlan_url, headers=RESTCONF_HEADERS,  timeout=10.0)
 
             response.raise_for_status()
             return {"detail": f"VLAN '{vlan_name}' successfully deleted from the switch."}
-    except Exception as e:
-            raise HTTPException(
-                status_code=500,
-                detail=str(e)
-            )
+        
+    except httpx.HTTPStatusError as e:
+        raise HTTPException(status_code=e.response.status_code, detail=str(e))  
 
 
 
 async def delete_vlan_description_by_name(vlan_name: str):
+
     url = f"{SONIC_BASE_URL}/restconf/data/sonic-vlan:sonic-vlan/VLAN/VLAN_LIST={vlan_name}/description"
-    try:
+    
+    exist = await check_Vlan_exist(vlan_name)
+    if not exist:
+        raise HTTPException(status_code=404, detail="Vlan name not found")
+    
+    try:    
         async with httpx.AsyncClient(verify=False) as client:
             response = await client.delete(url, headers=RESTCONF_HEADERS,  timeout=10.0)
 
             response.raise_for_status()
             return {"detail": f"VLAN '{vlan_name}' Description successfully deleted from the switch."}
-    except Exception as e:
-            raise HTTPException(
-                status_code=500,
-                detail=str(e)
-            )
+
+    except httpx.HTTPStatusError as e:
+        raise HTTPException(status_code=e.response.status_code, detail=str(e))
